Reärranged the Zones command concerning "at [x] [z]"/"at [x] [y] [z]"
Reärranged the Zones command, to put the "at [x] [y]"/"at [x] [y] [z]" pattern into a separate function referred to in the declaration of the command.

Changed the argument type of ZonesCommandHandler.getMutateZoneSubCommand(...) to accept the same functional interface as is used by Brigadier's .executes(...) method.

Renamed ZonesCommandHandler.getMutateZoneSubCommand to .getMutateZoneSubcommand.